+++ http_server/flask_blueprints/sensor_control_files/reports.py
@@ -17,20 +17,22 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 import time
+from threading import Thread
 from operations_modules import logger
 from operations_modules import file_locations
-from operations_modules.app_generic_functions import get_response_bg_colour, get_http_sensor_reading, \
-    get_file_content, check_for_port_in_address, get_ip_and_port_split
-from configuration_modules import app_config_access
 from operations_modules import app_cached_variables
+from configuration_modules import app_config_access
 from operations_modules import network_wifi
 from operations_modules import software_version
+from operations_modules.app_generic_functions import get_response_bg_colour, get_http_sensor_reading, \
+    get_file_content, check_for_port_in_address, get_ip_and_port_split
 from configuration_modules.config_primary import CreatePrimaryConfiguration
 from configuration_modules.config_installed_sensors import CreateInstalledSensorsConfiguration
 from configuration_modules.config_display import CreateDisplayConfiguration
 from configuration_modules.config_weather_underground import CreateWeatherUndergroundConfiguration
 from configuration_modules.config_luftdaten import CreateLuftdatenConfiguration
 from configuration_modules.config_open_sense_map import CreateOpenSenseMapConfiguration
+from sensor_modules.sensor_access import get_system_datetime
 
 running_with_root = app_cached_variables.running_with_root
 if software_version.old_version == software_version.version:
@@ -297,3 +299,51 @@ def get_online_report(ip_address, report_type="systems_report"):
     except Exception as error:
         log_msg = "Remote Sensor " + ip_address + " Failed providing " + str(report_type) + " Data: " + str(error)
         logger.network_logger.warning(log_msg)
+
+
+def get_sensor_control_report(address_list, report_type="systems_report"):
+    """
+    Returns a HTML report based on report_type and sensor addresses provided (IP or DNS addresses as a list).
+    Default: systems_report
+    """
+    config_report = app_config_access.sensor_control_config.radio_report_config
+    sensors_report = app_config_access.sensor_control_config.radio_report_test_sensors
+    latency_report = app_config_access.sensor_control_config.radio_report_sensors_latency
+    html_sensor_report_end = html_report_system_end
+
+    new_report = html_report_system_start
+    if report_type == config_report:
+        new_report = html_report_config_start
+        html_sensor_report_end = html_report_config_end
+    elif report_type == sensors_report:
+        new_report = html_report_sensors_test_start
+        html_sensor_report_end = html_report_sensors_test_end
+    elif report_type == latency_report:
+        new_report = html_report_sensors_latency_start
+        html_sensor_report_end = html_report_sensors_test_end
+    new_report = new_report.replace("{{ DateTime }}", get_system_datetime())
+
+    sensor_reports = []
+    threads = []
+    for address in address_list:
+        threads.append(Thread(target=get_online_report, args=[address, report_type]))
+    for thread in threads:
+        thread.start()
+    for thread in threads:
+        thread.join()
+
+    data_queue = app_cached_variables.data_queue
+    if report_type == config_report:
+        data_queue = app_cached_variables.data_queue2
+    if report_type == sensors_report or report_type == latency_report:
+        data_queue = app_cached_variables.data_queue3
+    while not data_queue.empty():
+        sensor_reports.append(data_queue.get())
+        data_queue.task_done()
+
+    sensor_reports.sort()
+    for report in sensor_reports:
+        new_report += str(report[1])
+    new_report += html_sensor_report_end
+    return new_report
+
